+++ Abilities/Rulebook/_rulebookBloodBath.cs
@@ -1,0 +1,27 @@
+﻿using InscryptionAPI;
+using DiskCardGame;
+using System.Collections;
+using UnityEngine;
+using Resources = WhistleWindLobotomyMod.Properties.Resources;
+
+namespace WhistleWindLobotomyMod
+{
+    public partial class WstlPlugin
+    {
+        private void Rulebook_BloodBath()
+        {
+            const string rulebookName = "Bloodbath";
+            const string rulebookDescription = "Transforms whenever another card is sacrificed, up to three times.";
+            const string dialogue = "femboy";
+            EntryBloodBath.ability = WstlUtils.CreateAbility<EntryBloodBath>(
+                Resources.sigilAbnormality,
+                rulebookName, rulebookDescription, dialogue, 0,
+                overrideModular: true).Id;
+        }
+    }
+    public class EntryBloodBath : AbilityBehaviour
+    {
+        public static Ability ability;
+        public override Ability Ability => ability;
+    }
+}