+++ Core/Patchers/CardPatcher.cs
@@ -8,14 +8,34 @@ using UnityEngine;
 
 namespace WhistleWindLobotomyMod
 {
-    public static class ForcedEmissions
+    public static class CardPatcher
     {
+        [HarmonyPatch(typeof(CardMergeSequencer), "GetValidCardsForHost")]
+        [HarmonyPostfix]
+        public static void RemoveFromValidCardsForHost(ref List<CardInfo> __result)
+        {
+            __result.RemoveAll((CardInfo x) => x.SpecialAbilities.Contains(NothingThere.specialAbility));
+        }
+
+        [HarmonyPatch(typeof(Deathtouch), "RespondsToDealDamage")]
+        [HarmonyPrefix]
+        public static bool ImmunetoDeathTouch(ref int amount, ref PlayableCard target)
+        {
+            bool whiteNightEvent = !target.HasAbility(TrueSaviour.ability) && !target.HasAbility(Apostle.ability) && !target.HasAbility(Confession.ability);
+            if (amount > 0 && target != null && !target.Dead)
+            {
+                return whiteNightEvent && !target.HasAbility(Ability.MadeOfStone);
+            }
+            return false;
+        }
+
         // Forces cards to render their emissive texture
         // Only for a select few cards
         // Let's not get too liberal with the 'make everything glowy' button
+        // Update: I got too liberal with the 'make everything glowy' button
         [HarmonyPatch(typeof(Card), "ApplyAppearanceBehaviours")]
         [HarmonyPostfix]
-        public static void Emissions(ref Card __instance)
+        public static void ForcedEmissions(ref Card __instance)
         {
             switch (__instance.Info.name.ToLowerInvariant())
             {
@@ -69,6 +89,15 @@ namespace WhistleWindLobotomyMod
                     __instance.StatsLayer.SetEmissionColor(GameColors.Instance.brightNearWhite);
                     __instance.RenderCard();
                     break;
+                case "wstl_hundredsgooddeeds":
+                    __instance.RenderInfo.forceEmissivePortrait = true;
+                    __instance.StatsLayer.SetEmissionColor(GameColors.Instance.brightNearWhite);
+                    __instance.RenderCard();
+                    break;
+                default: // for testing only
+                    __instance.RenderInfo.forceEmissivePortrait = true;
+                    __instance.RenderCard();
+                    break; /* */
             }
         }
     }