+++ Properties/Resources.Designer.cs
@@ -1043,6 +1043,16 @@ namespace WhistleWindLobotomyMod.Properties {
         /// <summary>
         ///   Looks up a localized resource of type System.Byte[].
         /// </summary>
+        public static byte[] meltingLove_emission {
+            get {
+                object obj = ResourceManager.GetObject("meltingLove_emission", resourceCulture);
+                return ((byte[])(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Byte[].
+        /// </summary>
         public static byte[] meltingLoveMinion {
             get {
                 object obj = ResourceManager.GetObject("meltingLoveMinion", resourceCulture);